FIFA 23 Workbook
Something that I put together after buying one of my favorite Video Games. This tool is especially helpful when trying to put together an effective team in the Ultimate Team mode. Noteworthy information about the dashboard is that these ratings do not incorporate live form and are based on the player ratings given when the game initially released. Hope you enjoy :)